backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
Flask==3.0.1
Flask-Cors==4.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
openai==1.13.3
packaging==23.2
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
replicate==0.24.0
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.1.0
Werkzeug==3.0.1
